requirements.txt
python-dotenv==1.0.1
requests>=2.31.0
Pillow>=10.3.0
openai==1.86.0
chainlit==2.5.5
chromadb==1.0.12
torch==2.6.0
torchvision==0.18.1
transformers>=4.37.2
ftfy>=6.1.1
regex>=2023.12.25
tqdm>=4.66.0
